feat: JSON-based Nix configuration (#115)
Added a migration to the daemon to shift the NixOS configuration to
`.json` based files instead of `.nix` so that the daemon can manage the
full configuration without resorting to the string parsing replacement
methods. Also fixed up some bugs left over from the previous PR. Still
need to go back and do the same type of change for k3s manifests.

+++ services/platform/daemon/host/stats.go
@@ -68,11 +68,11 @@ func getMemoryStats(stats *v1.SystemStats) error {
 		return err
 	}
 	stats.Memory = &v1.MemoryStats{
-		TotalBytes:     uint32(memory.Total),
-		UsedBytes:      uint32(memory.Used),
-		CachedBytes:    uint32(memory.Cached),
-		FreeBytes:      uint32(memory.Free),
-		AvailableBytes: uint32(memory.Available),
+		TotalBytes:     memory.Total,
+		UsedBytes:      memory.Used,
+		CachedBytes:    memory.Cached,
+		FreeBytes:      memory.Free,
+		AvailableBytes: memory.Available,
 	}
 	return nil
 }
@@ -85,8 +85,8 @@ func getDriveStats(stats *v1.SystemStats, mounts []string) error {
 			return err
 		}
 		// Available blocks * size per block = available space in bytes
-		free := uint32(stat.Bavail * uint64(stat.Bsize))
-		total := uint32(stat.Blocks) * uint32(stat.Bsize)
+		free := stat.Bavail * uint64(stat.Bsize)
+		total := stat.Blocks * uint64(stat.Bsize)
 		stats.Drives[index] = &v1.DriveStats{
 			MountPoint: mountPoint,
 			TotalBytes: total,
